% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables_names.R
\name{get_variables_names}
\alias{get_variables_names}
\title{Obtain variable names}
\usage{
get_variables_names()
}
\value{
A vector containing the unique variables names in all regressions
}
\description{
Obtain variable names
}
\examples{
NULL
}
