% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_change_covnames.R
\name{mod_change_covnames_server}
\alias{mod_change_covnames_server}
\title{change_covnames Server Function}
\usage{
mod_change_covnames_server(input, output, session)
}
\value{
This returns three outputs:
\itemize{
  \item ancient - A vector containing all values in \code{textInput} in the column "current names".
  \item new - A vector containing all values in \code{textInput} in the column "new names".
  \item button - The value taken by the button "Apply". This is useful in \code{server} part to save the values stored in "new". Indeed, this is the only way to order the two buttons (the one to launch the modal and the one to apply the changes).
}
}
\description{
change_covnames Server Function
}
