% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_supported.R
\name{models_supported_stargazer}
\alias{models_supported_stargazer}
\title{List of models supported by {stargazer}}
\usage{
models_supported_stargazer()
}
\value{
All the models supported by {stargazer}, see the list here: https://rdrr.io/cran/stargazer/man/stargazer_models.html
}
\description{
List of models supported by {stargazer}
}
\examples{
NULL
}
