% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_options.R
\name{result_options}
\alias{result_options}
\title{Result options
Function that details the content of the "Result options" dropdown}
\usage{
result_options()
}
\value{
The part of the Shiny UI that details the options for the results in the {\code{stargazer}} table.
}
\description{
Result options
Function that details the content of the "Result options" dropdown
}
\examples{
NULL
}
